+++ adapters/unsupported/scryer.pl
@@ -1,0 +1,650 @@
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  Adapter file for Scryer Prolog
+%  Last updated on April 3, 2019
+%
+%  This file is part of Logtalk <https://logtalk.org/>
+%  Copyright 1998-2019 Paulo Moura <pmoura@logtalk.org>
+%
+%  Licensed under the Apache License, Version 2.0 (the "License");
+%  you may not use this file except in compliance with the License.
+%  You may obtain a copy of the License at
+%
+%      http://www.apache.org/licenses/LICENSE-2.0
+%
+%  Unless required by applicable law or agreed to in writing, software
+%  distributed under the License is distributed on an "AS IS" BASIS,
+%  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+%  See the License for the specific language governing permissions and
+%  limitations under the License.
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  ISO Prolog Standard predicates that we must define because they are
+%  not built-in
+%
+%  add a clause for '$lgt_iso_predicate'/1 declaring each ISO predicate that
+%  we must define; there must be at least one clause for this predicate
+%  whose call should fail if we don't define any ISO predicates
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_iso_predicate'(?callable).
+%
+% table of missing ISO predicates which are defined in this file
+
+'$lgt_iso_predicate'(_) :-
+	fail.
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  de facto standard Prolog predicates that might be missing
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% between(+integer, +integer, ?integer)
+
+
+% findall(?term, +callable, ?list, +list)
+
+
+% forall(+callable, +callable)
+
+forall(Generate, Test) :-
+	\+ (Generate, \+ Test).
+
+
+% format(+stream_or_alias, +character_code_list_or_atom, +list) -- built-in ????
+
+format(Stream, Format, Arguments) :-
+	?????
+
+
+% format(+character_code_list_or_atom, +list) -- built-in ????
+
+format(Format, Arguments) :-
+	?????
+
+
+% numbervars(?term, +integer, ?integer)
+
+:- use_module(library(terms), [numbervars/3]).
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  predicate properties
+%
+%  this predicate must return at least static, dynamic, and built_in
+%  properties for an existing predicate (and ideally meta_predicate/1
+%  properties for built-in predicates and library predicates)
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_predicate_property'(+callable, ?predicate_property)
+
+'$lgt_predicate_property'(Pred, Prop) :-
+	?????
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  meta-predicates
+%
+%  the setup_call_cleanup/3 meta-predicate is only strictly required if
+%  the backend Prolog compiler supports a compatible multi-threading
+%  implementation but the developer tools can also take advantage of it
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% setup_call_cleanup(+callable, +callable, +callable)
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  Prolog non-standard built-in meta-predicates and meta-directives
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_prolog_meta_predicate'(@callable, ?callable, ?atom)
+%
+% table of meta-predicate patterns for proprietary built-in predicates;
+% the third argument, which must be either "predicate" or "control_construct",
+% is used to guide the compilation of these meta-predicates in debug mode
+
+'$lgt_prolog_meta_predicate'(setup_call_cleanup(_, _, _), setup_call_cleanup(0, 0, 0), predicate).
+'$lgt_prolog_meta_predicate'(call_cleanup(_, _), call_cleanup(0, 0), predicate).
+'$lgt_prolog_meta_predicate'(freeze(_, _), freeze(*, 0), predicate).
+'$lgt_prolog_meta_predicate'(call_with_inference_limit(_, _, _), call_with_inference_limit(0, *, *), predicate).
+
+
+% '$lgt_prolog_meta_directive'(@callable, -callable)
+
+'$lgt_prolog_meta_directive'(_, _) :-
+	fail.
+
+
+% '$lgt_prolog_to_logtalk_meta_argument_specifier_hook'(@nonvar, -atom)
+
+'$lgt_prolog_to_logtalk_meta_argument_specifier_hook'(_, _) :-
+	fail.
+
+
+% '$lgt_candidate_tautology_or_falsehood_goal_hook'(@callable)
+%
+% valid candidates are proprietary built-in predicates with
+% no side-effects when called with ground arguments
+
+'$lgt_candidate_tautology_or_falsehood_goal_hook'(_) :-
+	fail.
+
+
+% '$lgt_prolog_database_predicate'(@callable)
+
+'$lgt_prolog_database_predicate'(_) :-
+	fail.
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  file name extension predicates
+%
+%  these extensions are used by Logtalk load/compile predicates
+%
+%  you may want to change the extension for the intermediate files
+%  generated by the Logtalk compiler ("object" files) to match the
+%  extension expected by default by your Prolog compiler
+%
+%  there should only a single extension defined for object files but
+%  but multiple extensions can be defined for Logtalk and Prolog source
+%  files and for backend specific temporary files
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_file_extension'(?atom, ?atom)
+
+'$lgt_file_extension'(logtalk, '.lgt').
+'$lgt_file_extension'(logtalk, '.logtalk').
+'$lgt_file_extension'(object, '.pl').
+'$lgt_file_extension'(prolog, '.pl').
+'$lgt_file_extension'(prolog, '.prolog').
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  backend Prolog compiler features
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_prolog_feature'(?atom, ?atom)
+%
+% backend Prolog compiler supported features (that are compatible with Logtalk)
+
+'$lgt_prolog_feature'(prolog_dialect, scryer).
+'$lgt_prolog_feature'(prolog_version, v(Major, Minor, Path)) :-
+	?????
+'$lgt_prolog_feature'(prolog_compatible_version, '@>='(v(Major, Minor, Path))).
+'$lgt_prolog_feature'(prolog_conformance, lax).	% either lax or strict
+
+'$lgt_prolog_feature'(encoding_directive, unsupported).
+'$lgt_prolog_feature'(tabling, unsupported).
+'$lgt_prolog_feature'(threads, unsupported).
+'$lgt_prolog_feature'(modules, supported).
+'$lgt_prolog_feature'(coinduction, unsupported).
+'$lgt_prolog_feature'(unicode, unsupported).
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  default flag values
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_default_flag'(?atom, ?atom)
+%
+% default values for all flags
+
+% startup flags:
+'$lgt_default_flag'(settings_file, allow).
+% lint compilation flags:
+'$lgt_default_flag'(unknown_entities, warning).
+'$lgt_default_flag'(unknown_predicates, warning).
+'$lgt_default_flag'(undefined_predicates, warning).
+'$lgt_default_flag'(singleton_variables, warning).
+'$lgt_default_flag'(steadfastness, silent).
+'$lgt_default_flag'(naming, silent).
+'$lgt_default_flag'(portability, silent).
+'$lgt_default_flag'(redefined_built_ins, silent).
+'$lgt_default_flag'(deprecated, warning).
+'$lgt_default_flag'(missing_directives, warning).
+'$lgt_default_flag'(duplicated_directives, warning).
+'$lgt_default_flag'(trivial_goal_fails, warning).
+'$lgt_default_flag'(always_true_or_false_goals, warning).
+'$lgt_default_flag'(lambda_variables, warning).
+'$lgt_default_flag'(suspicious_calls, warning).
+'$lgt_default_flag'(underscore_variables, singletons).
+% optional features compilation flags:
+'$lgt_default_flag'(complements, deny).
+'$lgt_default_flag'(dynamic_declarations, deny).
+'$lgt_default_flag'(events, deny).
+'$lgt_default_flag'(context_switching_calls, allow).
+% other compilation flags:
+'$lgt_default_flag'(scratch_directory, './lgt_tmp/').
+'$lgt_default_flag'(report, on).
+'$lgt_default_flag'(clean, on).
+'$lgt_default_flag'(code_prefix, '$').
+'$lgt_default_flag'(optimize, off).
+'$lgt_default_flag'(source_data, on).
+'$lgt_default_flag'(reload, changed).
+'$lgt_default_flag'(debug, off).
+% Prolog compiler and loader flags:
+'$lgt_default_flag'(prolog_compiler, []).
+'$lgt_default_flag'(prolog_loader, []).
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  operating-system access predicates
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_prolog_os_file_name'(+atom, -atom)
+% '$lgt_prolog_os_file_name'(-atom, +atom)
+%
+% converts between Prolog internal file paths and operating-system paths
+
+'$lgt_prolog_os_file_name'(PrologPath, OSPath) :-
+	?????
+
+
+% '$lgt_expand_path'(+nonvar, -atom)
+%
+% expands a file path to a full path
+
+'$lgt_expand_path'(Path, ExpandedPath) :-
+	?????
+
+
+% '$lgt_file_exists'(+atom)
+%
+% checks if a file exists
+
+'$lgt_file_exists'(File) :-
+	?????
+
+
+% '$lgt_delete_file'(+atom)
+%
+% deletes a file
+
+'$lgt_delete_file'(File) :-
+	?????
+
+
+% '$lgt_directory_exists'(+atom)
+%
+% checks if a directory exists
+
+'$lgt_directory_exists'(Directory) :-
+	?????
+
+
+% '$lgt_current_directory'(-atom)
+%
+% gets current working directory
+
+'$lgt_current_directory'(Directory) :-
+	?????
+
+
+% '$lgt_change_directory'(+atom)
+%
+% changes current working directory
+
+'$lgt_change_directory'(Directory) :-
+	?????
+
+
+% '$lgt_make_directory'(+atom)
+%
+% makes a new directory; succeeds if the directory already exists
+
+'$lgt_make_directory'(Directory) :-
+	?????
+
+
+% '$lgt_directory_hash_as_atom'(+atom, -atom)
+%
+% returns the directory hash as an atom
+
+'$lgt_directory_hash_as_atom'(_, '').
+
+
+% '$lgt_compile_prolog_code'(+atom, +atom, +list)
+%
+% compile to disk a Prolog file, resulting from a
+% Logtalk source file, given a list of flags
+
+'$lgt_compile_prolog_code'(File, Source, Options) :-
+	?????.
+
+
+% '$lgt_load_prolog_code'(+atom, +atom, +list)
+%
+% compile and load a Prolog file, resulting from a
+% Logtalk source file, given a list of flags
+
+'$lgt_load_prolog_code'(File, Source, Options) :-
+	?????
+
+
+% '$lgt_file_modification_time'(+atom, -nonvar)
+%
+% gets a file modification time, assumed to be an opaque term but comparable
+
+'$lgt_file_modification_time'(File, Time) :-
+	?????
+
+
+% '$lgt_environment_variable'(?atom, ?atom)
+%
+% access to operating-system environment variables
+
+'$lgt_environment_variable'(Variable, Value) :-
+	?????
+
+
+% '$lgt_decompose_file_name'(+atom, ?atom, ?atom, ?atom)
+%
+% decomposes a file path in its components; the directory must always end
+% with a slash; the extension must start with a "." when defined and must
+% be the empty atom when it does not exist
+
+'$lgt_decompose_file_name'(File, Directory, Name, Extension) :-
+	?????
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  getting stream current line number
+%  (needed for improved compiler error messages)
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_stream_current_line_number'(@stream, -integer)
+
+'$lgt_stream_current_line_number'(Stream, Line) :-
+	?????
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  abstraction of the standard open/4 and close/1 predicates for dealing
+%  with the alias/1 option in old non-standard compliant systems
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_open'(+atom, +atom, -stream, @list)
+
+'$lgt_open'(File, Mode, Stream, Options) :-
+	?????
+
+
+% '$lgt_close'(@stream)
+
+'$lgt_close'(Stream) :-
+	?????
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  customized version of the read_term/3 predicate for returning the term
+%  position (start and end lines; needed for improved error messages) and
+%  the variable names (ideally using the standard variable_names/1 option)
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_read_term'(@stream, -term, +list, -position, -list)
+
+'$lgt_read_term'(Stream, Term, Options, LineBegin-LineEnd, Variables) :-
+	?????
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  Prolog dialect specific term and goal expansion
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_prolog_term_expansion'(@callable, -callable)
+
+'$lgt_prolog_term_expansion'(_, _) :-
+	fail.
+
+
+% '$lgt_prolog_goal_expansion'(@callable, -callable)
+
+'$lgt_prolog_goal_expansion'(_, _) :-
+	fail.
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  converts between Prolog stream encoding names and XML encoding names
+%  (only necessary for Prolog compilers supporting different text encodings;
+%  for others simply provide a dummy definition that always fail)
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_logtalk_prolog_encoding'(?atom, ?atom)
+
+'$lgt_logtalk_prolog_encoding'(_, _, _) :-
+	fail.
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  lambda expressions support predicates
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_copy_term_without_constraints'(@term, ?term)
+
+'$lgt_copy_term_without_constraints'(Term, Copy) :-
+	copy_term(Term, Copy).
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  hooks predicates for writing and asserting compiled entity terms
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_write_compiled_term'(@stream, @callable, +atom, +atom, +integer)
+%
+% the third argument is the term type: runtime (internal runtime clause),
+% user (compiled user-defined term), or aux (auxiliary clause resulting
+% e.g. from term-expansion)
+
+'$lgt_write_compiled_term'(Stream, Term, _Kind, _Path, _Line) :-
+	write_canonical(Stream, Term),
+	write(Stream, '.\n').
+
+
+% '$lgt_assertz_entity_clause'(@clause, +atom)
+
+'$lgt_assertz_entity_clause'(Clause, _Kind) :-
+	assertz(Clause).
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  error term normalization (when exception terms don't follow the ISO
+%  Prolog standard)
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_normalize_error_term'(@callable, -callable)
+
+'$lgt_normalize_error_term'(
+	Error,
+	Error
+).
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  message token printing
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+%:- multifile('$logtalk#0.print_message_token#4'/5).
+%:- dynamic('$logtalk#0.print_message_token#4'/5).
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  term hashing (not currently used in the compiler/runtime)
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% term_hash(@callable, +integer, +integer, -integer)
+
+term_hash(_, _, _, _) :-
+	fail.
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  string built-in type
+%
+%  define these predicates to trivially fail if no string type is available
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_string'(@term)
+
+'$lgt_string'(Term) :-
+	string(Term).
+
+
+% '$lgt_string_codes'(+string, -list(codes))
+% '$lgt_string_codes'(-string, +list(codes))
+
+'$lgt_string_codes'(String, Codes) :-
+	?????
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  module qualification to be used when calling Prolog meta-predicates
+%  with meta-arguments that are calls to object or category predicates
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+% '$lgt_user_module_qualification'(@callable, -callable)
+
+'$lgt_user_module_qualification'(Goal, Goal).
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  shortcuts to the Logtalk built-in predicates logtalk_load/1 and
+%  logtalk_make/1
+%
+%  defined in the adapter files to make it easier to comment them out in case
+%  of conflict with some Prolog native feature; they require conformance with
+%  the ISO Prolog standard regarding the definition of the {}/1 syntax
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
+
+{*} :-
+	!,
+	logtalk_make(all).
+{!} :-
+	!,
+	logtalk_make(clean).
+{?} :-
+	!,
+	logtalk_make(check).
+{@} :-
+	!,
+	logtalk_make(circular).
+{#} :-
+	!,
+	logtalk_make(documentation).
+{+d} :-
+	!,
+	logtalk_make(debug).
+{+n} :-
+	!,
+	logtalk_make(normal).
+{+o} :-
+	!,
+	logtalk_make(optimal).
+{$} :-
+	!,
+	logtalk_make(caches).
+
+{File, Files} :-
+	!,
+	logtalk_load(File),
+	{Files}.
+{File} :-
+	logtalk_load(File).
+
+
+
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%
+%  end!
+%
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%